pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verbatim-rag"
version = "0.1.0"
description = "A minimalistic RAG system that prevents hallucination by ensuring all generated content is explicitly derived from source documents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Adam Kovacs", email = "kovacs@krlabs.eu"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.3.0",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.1",
]

[project.urls]
"Homepage" = "https://github.com/krlabsorg/verbatim-rag"
"Bug Tracker" = "https://github.com/krlabsorg/verbatim-rag/issues"

[project.scripts]
verbatim-rag = "verbatim_rag.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["verbatim_rag"] 